+++ crates/plugin_runtime/src/lib.rs
@@ -59,6 +59,9 @@ impl Plugin for FarmPluginRuntime {
   }
 
   fn config(&self, config: &mut Config) -> farmfe_core::error::Result<Option<()>> {
+    if config.output.target_env.is_library() {
+      return Ok(None);
+    }
     // runtime package entry file
     if !config.runtime.path.is_empty() {
       config.input.insert(
@@ -265,7 +268,9 @@ impl Plugin for FarmPluginRuntime {
     context: &Arc<CompilationContext>,
     _hook_context: &PluginHookContext,
   ) -> farmfe_core::error::Result<Option<ResourcePotMetaData>> {
-    if matches!(resource_pot.resource_pot_type, ResourcePotType::Js) {
+    if !context.config.output.target_env.is_library()
+      && matches!(resource_pot.resource_pot_type, ResourcePotType::Js)
+    {
       let async_modules = self.get_async_modules(context);
       let async_modules = async_modules.downcast_ref::<HashSet<ModuleId>>().unwrap();
       let module_graph = context.module_graph.read();
@@ -439,6 +444,10 @@ impl Plugin for FarmPluginRuntime {
     param: &mut PluginFinalizeResourcesHookParams,
     context: &Arc<CompilationContext>,
   ) -> farmfe_core::error::Result<Option<()>> {
+    if context.config.output.target_env.is_library() {
+      return Ok(None);
+    }
+
     let async_modules = self.get_async_modules(context);
     let async_modules = async_modules.downcast_ref::<HashSet<ModuleId>>().unwrap();
     handle_entry_resources::handle_entry_resources(param.resources_map, context, async_modules);