+++ packages/core/src/types/swc-config.ts
@@ -1,4 +1,3 @@
-
 export interface ScriptParseConfig {
   esConfig?: {
     jsx?: boolean;
@@ -81,9 +80,7 @@ export interface SwcPresetEnvOptions {
   forceAllTransforms?: boolean;
 }
 
-interface TerserEcmaVersion {
-
-}
+interface TerserEcmaVersion {}
 
 interface TerserManglePropertiesOptions {}
 
@@ -250,4 +247,9 @@ export interface JsMinifyOptions {
    * @default 'minify-module'
    */
   mode?: 'minify-module' | 'minify-resource-pot';
+
+  /**
+   * @default true
+   */
+  moduleDecls?: boolean;
 }