+++ crates/plugin_bundle/CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5](https://github.com/farm-fe/farm/compare/farmfe_plugin_bundle-v0.0.4...farmfe_plugin_bundle-v0.0.5) - 2024-07-25
+
+### Added
+- support targetEnv library library-browser/node ([#1656](https://github.com/farm-fe/farm/pull/1656))
+
+## [0.0.4](https://github.com/farm-fe/farm/compare/farmfe_plugin_bundle-v0.0.3...farmfe_plugin_bundle-v0.0.4) - 2024-07-19
+
+### Added
+- support single library bundle ([#1640](https://github.com/farm-fe/farm/pull/1640))
+
+## [0.0.3](https://github.com/farm-fe/farm/compare/farmfe_plugin_bundle-v0.0.2...farmfe_plugin_bundle-v0.0.3) - 2024-07-15
+
+### Other
+- updated the following local packages: farmfe_core
+
 ## [0.0.2](https://github.com/farm-fe/farm/compare/farmfe_plugin_bundle-v0.0.1...farmfe_plugin_bundle-v0.0.2) - 2024-07-11
 
 ### Other